Move keywords to a clear OpKeyword* lexems:
I'm leaving DoBlock currently the same, but the rest will be under
OpKeyword*Expr to express the syntax for those particular ones.

That means that in the future, DoBlock will turn into OpKeywordDoExpr.

It might not mean the same for flows. I don't see OpKeywordForExpr to
be a comfortable expression name.