Add a hell of a lot of keywords:
I've had a bit of a confusing time here.

These keywords should receive values (Value) but because it's Perl, every Value
can be an Expression. The problem is having two Expressions one after the other.

    OpKeywordAccept Expression Expression

I think this is just really odd to parse, because a single Expression can return
multiple values ( "@{$foo}", for example ). I'm not sure how Perl handles it other
than being very confused:

    # Fails to parse with Expression of deref
    $ perl -c -Mstrict -W -E'my $foo = [ 1, 2 ]; accept( @{ $foo } );'
    Not enough arguments for accept at -e line 1, near "} )"
    -e had compilation errors.

    # Succeeds to parse with Expression of a function call
    $ perl -c -Mstrict -W -E'sub foo () {1}; accept( foo(), foo() );'
    -e syntax OK

To handle this what I picked was using an Expression when it's a single argument
and using a set of Values when it's more than one argument. Maybe it's good enough.
Once it breaks code, we could review if we want to expand what can be used instead of
just a full Expression. Maybe we *do* want multiple Expressions... Thoughts?

The following keywords are added:

* accept
* alarm
* atan2
* bind
* binmode
* bless
* break
* caller
* chdir
* chmod
* chomp
* chop
* chown
* chr
* chroot
* close
* closedir
* connect
* continue
* cos
* crypt
* dbmclose
* dbmopen
* defined
* delete
* die
* do
* dump
* each
* eof
* eval (only block-based, no expression)
* evalbytes
* exec
* exists
* exit
* exp
* fc
* fcntl
* fileno
* flock
* fork
* getc
* getlogin
* getpeername
* getpgrp
* getppid
* getpriority
* getpwnam
* getgrnam
* gethostbyname
* getnetbyname
* getprotobyname
* getservbyport
* getpwent
* getgrent
* gethostent
* getnetent
* getprotoent
* getservent
* setpwent
* setgrent
* sethostent
* setnetent
* setprotoent
* setservent
* endpwent
* endgrent
* endhostent
* endnetent
* endprotoent
* endservent
* getsockname
* getsockopt