Support subroutines with names of Phase statements:
To support `sub BEGIN {}` et. al., we need to do the same crazy trick
we've done for quote-like functions.

Because we can't provide a negation for an entire string (not to mention
the lack of '&&' capability in rules), we need to explcitily state
all possibilities.

Here's the logic:

* B - ok
* B Non-E - ok
* B E Non-G - ok
* B E G Non-I - ok
* B E G I Non-N - ok

This means that "BEGIN" will not be allowed.

For every letter check of uppercase, we need our range to also support
all lowercacse, so "BEGin" will be support while "BEGIN" won't.

Same for the opposite. "qq" should be identified, but "QQ" shouldn't.