Add some tests for SubCall, with a TODO:
While fixing up tests, I found that we don't tell the difference between
Ident used for classes and ones used for subroutine names.

They're mostly similar, but the following is different:

1. Foo::->bar()  # method call, works
2. Foo->bar()    # same, ambiguous to perl, but not to guacamole
3. Foo::bar()    # Function call, Foo class, bar function name
4. Foo::bar::()  # Error!

But because we use Ident for all of these, #1 and #4 are getting confused.

We can fix it. It's not a top priority.