Support all Perl's special variables:
This is done by separating them from the definition of Indent for
the VarScalar.

We could theoretically put them under VarScalar, VarArray, and
VarHash, but I'm not really sure how to hack that in.

It might be a good argument to keep it as a separate value type.
In that case, I don't know if GlobalVariable is a good name for it.