Make grammar externally accessible, lazily-built but stateful:
This is useful so you can prod and adjust the grammar before the
grammar object is created.

I hate that we use a class method without an instance or allow
clearing or accessing the grammar object once we build it, but at
the very least, we now allow the string to be accessed and updated
before the grammar is created. We also cache it now with a stateful
variable.